Refactor LED pattern handling (#89)
