+++ vendordeps/PathplannerLib-2025.2.6.json
@@ -1,7 +1,7 @@
 {
-    "fileName": "PathplannerLib-2025.2.5.json",
+    "fileName": "PathplannerLib-2025.2.6.json",
     "name": "PathplannerLib",
-    "version": "2025.2.5",
+    "version": "2025.2.6",
     "uuid": "1b42324f-17c6-4875-8e77-1c312bc8c786",
     "frcYear": "2025",
     "mavenUrls": [
@@ -12,7 +12,7 @@
         {
             "groupId": "com.pathplanner.lib",
             "artifactId": "PathplannerLib-java",
-            "version": "2025.2.5"
+            "version": "2025.2.6"
         }
     ],
     "jniDependencies": [],
@@ -20,7 +20,7 @@
         {
             "groupId": "com.pathplanner.lib",
             "artifactId": "PathplannerLib-cpp",
-            "version": "2025.2.5",
+            "version": "2025.2.6",
             "libName": "PathplannerLib",
             "headerClassifier": "headers",
             "sharedLibrary": false,
